pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xva-core"
version = "1.0.0"
description = "Educational xVA calculation engine with SA-CCR"
authors = [{name = "Lucas", email = "lucas@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.28",
    "plotly>=5.17",
    "openpyxl>=3.1",
]
dev = [
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.7",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "tox>=4.0",
    "pre-commit>=3.5",
    "jupyter>=1.0",
    "ipykernel>=6.26",
    "types-PyYAML>=6.0",
]
all = ["xva-core[ui,dev]"]

[project.scripts]
xva-demo = "examples.run_demo:main"

[project.urls]
Homepage = "https://github.com/user/xva-project"
Documentation = "https://xva-project.readthedocs.io"
Repository = "https://github.com/user/xva-project"

[tool.setuptools.packages.find]
where = ["."]
include = ["xva_core*", "xva_app*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["D"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["scipy.*", "plotly.*", "streamlit.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=xva_core",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
    "-v"
]

[tool.coverage.run]
source = ["xva_core"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
